--- Projects/Core/Analytics/Project.swift
@@ -1,22 +1,0 @@
-//
-//  Project.swift
-//  ProjectDescriptionHelpers
-//
-//  Created by kyuchul on 11/1/24.
-//
-
-@preconcurrency import ProjectDescription
-import ProjectDescriptionHelpers
-import DependencyPlugin
-
-let project = Project.make(
-    name: ModulePath.Core.Analytics.rawValue,
-    targets: [
-        .core(implements: .Analytics, factory: .init(
-            dependencies: [
-                .shared,
-                .shared(implements: .ThirdParty)
-            ]
-        ))
-    ]
-)
