+++ Projects/Core/Clients/AnalyticsClient/Sources/Base/AnalyticsEventName.swift
