+++ Projects/Core/Clients/AnalyticsClient/Sources/Base/AnalyticsScreenName.swift
