+++ Projects/Core/Clients/AnalyticsClient/Sources/Base/AnalyticsyLogEvent.swift
