+++ Projects/Core/Clients/AnalyticsClient/Sources/Client/AnalyticsClient.swift
@@ -16,16 +16,3 @@ public struct AnalyticsClient {
   public var logEvent: @Sendable (_ event: AnalyticsLogEvent) -> Void
   public var setUserId: @Sendable (_ userID: String?) -> Void
 }
-
-extension AnalyticsClient: DependencyKey {
-  public static var liveValue: AnalyticsClient {
-    return Self(
-      logEvent: { event in
-        AnalyticsManager.shared.logEvent(event)
-      }, 
-      setUserId: { userID in
-        AnalyticsManager.shared.setUserId(userID)
-      }
-    )
-  }
-}
