+++ Projects/Feature/Scene/Auth/Login/LoginFeature.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-import Analytics
+import AnalyticsClient
 import AuthClient
 import BKModel
 import FolderClient
