+++ Projects/Feature/Scene/Auth/Onboarding/OnboardingFlow/OnboardingFlowFeature.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-import Analytics
+import AnalyticsClient
 import Services
 
 import ComposableArchitecture
