+++ Projects/Feature/Scene/Auth/Onboarding/OnboardingInfo/OnboardingSubject/OnboardingSubjectFeature.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-import Analytics
+import AnalyticsClient
 import Services
 import UserClient
