+++ Projects/Feature/Scene/SaveLink/SaveLinkFeature.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-import Analytics
+import AnalyticsClient
 import LinkClient
 import Services
